perf(app): optimize cells rendering
