+++ packages/app/src/game/components/field-cell-text/field-cell-text.styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+export const FieldCellTextStyles = StyleSheet.create({
+    textActive: {
+        fontWeight: 'bold'
+    }
+});
