+++ packages/app/src/game/components/field-cell/field-cell.styles.ts
@@ -1,10 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import {
-    CellCandidateHorizontalOffsetConstant,
-    CellCandidateVerticalOffsetConstant,
-    CellSizeConstant
-} from '../constants/dimensions.contant';
+import { CellSizeConstant } from '../constants/dimensions.contant';
 
 export const FieldCellStyles = StyleSheet.create({
     container: {
@@ -33,47 +29,5 @@ export const FieldCellStyles = StyleSheet.create({
     },
     lastRow: {
         borderBottomWidth: 1
-    },
-    textActive: {
-        fontWeight: 'bold'
-    },
-    textCandidate: {
-        position: 'absolute'
-    },
-    textCandidatePosition1: {
-        left: CellCandidateHorizontalOffsetConstant,
-        top: CellCandidateVerticalOffsetConstant
-    },
-    textCandidatePosition2: {
-        left: CellCandidateHorizontalOffsetConstant + CellSizeConstant / 3,
-        top: CellCandidateVerticalOffsetConstant
-    },
-    textCandidatePosition3: {
-        right: CellCandidateHorizontalOffsetConstant,
-        top: CellCandidateVerticalOffsetConstant
-    },
-    textCandidatePosition4: {
-        left: CellCandidateHorizontalOffsetConstant,
-        top: CellSizeConstant / 3
-    },
-    textCandidatePosition5: {
-        left: CellCandidateHorizontalOffsetConstant + CellSizeConstant / 3,
-        top: CellSizeConstant / 3
-    },
-    textCandidatePosition6: {
-        right: CellCandidateHorizontalOffsetConstant,
-        top: CellSizeConstant / 3
-    },
-    textCandidatePosition7: {
-        bottom: CellCandidateVerticalOffsetConstant,
-        left: CellCandidateHorizontalOffsetConstant
-    },
-    textCandidatePosition8: {
-        bottom: CellCandidateVerticalOffsetConstant,
-        left: CellCandidateHorizontalOffsetConstant + CellSizeConstant / 3
-    },
-    textCandidatePosition9: {
-        bottom: CellCandidateVerticalOffsetConstant,
-        right: CellCandidateHorizontalOffsetConstant
     }
 });
