+++ packages/app/src/game/components/field-cell/field-cell.tsx
@@ -6,26 +6,18 @@ import { type OnEventFn, cs } from '@rnw-community/shared';
 
 import { animationDurationConstant } from '../../../@generic/constants/animation.constant';
 import { useAppSelector } from '../../../@generic/hooks/use-app-selector.hook';
-import { settingsFontSizeMultiplierSelector, settingsKeySelector } from '../../../settings/store/settings.selectors';
+import { settingsKeySelector } from '../../../settings/store/settings.selectors';
 import { ThemeContext } from '../../../theme/context/theme.context';
 import { GameContext } from '../../context/game.context';
-import { CellCandidateFontSizeConstant, CellCandidateMaxFontSizeConstant, CellFontSizeConstant } from '../constants/dimensions.contant';
 
 import { FieldCellSelectors as selectors } from './field-cell.selectors';
 import { FieldCellStyles as styles } from './field-cell.styles';
 
 import type { BWDarkTheme } from '../../../theme/themes/bw.theme';
 import type { CellInterface } from '@suuudokuuu/generator';
+import type { ReactNode} from 'react';
 
 const ReanimatedPressable = Reanimated.createAnimatedComponent(Pressable);
-
-const getText = (isActive: boolean, isEmpty: boolean, hasCandidates: boolean, cell: CellInterface): string => {
-    if (isEmpty) {
-        return isActive && !hasCandidates ? '•' : '';
-    }
-
-    return cell.value.toString();
-};
 
 const getCellBgColor = (
     theme: typeof BWDarkTheme,
@@ -53,29 +45,6 @@ const getCellBgColor = (
     return theme.colors.white;
 };
 
-const getCellTextColor = (
-    theme: typeof BWDarkTheme,
-    isActive: boolean,
-    isEmpty: boolean,
-    isHighlighted: boolean,
-    isActiveValue: boolean,
-    showAreas: boolean,
-    showIdenticalNumbers: boolean
-    // eslint-disable-next-line @typescript-eslint/max-params
-) => {
-    if (isActive) {
-        return theme.colors.cell.activeText;
-    } else if (isActiveValue && showIdenticalNumbers) {
-        return theme.colors.cell.activeValueText;
-    } else if (isHighlighted && showAreas) {
-        return theme.colors.cell.highlightedText;
-    } else if (isEmpty) {
-        return theme.colors.cell.emptyValueText;
-    }
-
-    return theme.colors.black;
-};
-
 const getCellSelector = (props: Props): selectors => {
     if (props.isActive) {
         return selectors.Active;
@@ -88,60 +57,29 @@ const getCellSelector = (props: Props): selectors => {
     return selectors.Root;
 };
 
-const getCandidateTextStyles = (theme: typeof BWDarkTheme, candidate: number, fontSizeMultiplier: number) => {
-    const textCandidatePositionStyles = {
-        1: styles.textCandidatePosition1,
-        2: styles.textCandidatePosition2,
-        3: styles.textCandidatePosition3,
-        4: styles.textCandidatePosition4,
-        5: styles.textCandidatePosition5,
-        6: styles.textCandidatePosition6,
-        7: styles.textCandidatePosition7,
-        8: styles.textCandidatePosition8,
-        9: styles.textCandidatePosition9
-    };
-
-    const textCandidateStyle = textCandidatePositionStyles[candidate as keyof typeof textCandidatePositionStyles];
-
-    return [
-        styles.textCandidate,
-        {
-            fontSize: Math.min(CellCandidateFontSizeConstant * fontSizeMultiplier, CellCandidateMaxFontSizeConstant),
-            color: theme.colors.cell.candidate
-        },
-        textCandidateStyle
-    ];
-};
-
 const animationConfig = { duration: animationDurationConstant };
 
 interface Props {
     readonly cell: CellInterface;
     readonly onSelect: OnEventFn<CellInterface | undefined>;
     readonly isActive: boolean;
+    readonly isEmpty: boolean;
     readonly isActiveValue: boolean;
     readonly isHighlighted: boolean;
     readonly isWrong: boolean;
-    readonly hasCandidates: boolean;
-    readonly hasAnimation: boolean;
-    readonly textAnimatedStyle: ReturnType<typeof useAnimatedStyle>;
+    readonly children?: ReactNode;
 }
 
-// TODO: Components is overloaded - split into smaller components
-// eslint-disable-next-line max-statements
 export const FieldCell = (props: Props) => {
-    const { cell, onSelect, isActive, isActiveValue, isHighlighted, isWrong, hasCandidates, hasAnimation, textAnimatedStyle } = props;
+    const { cell, onSelect, isActive, isActiveValue, isHighlighted, isWrong, isEmpty, children } = props;
 
     const { sudoku } = use(GameContext);
     const { theme } = use(ThemeContext);
 
-    const hasTextAnimation = useAppSelector(settingsKeySelector('showComboAnimation'));
     const showAreas = useAppSelector(settingsKeySelector('showAreas'));
     const showIdenticalNumbers = useAppSelector(settingsKeySelector('showIdenticalNumbers'));
     const showFilledNumbers = useAppSelector(settingsKeySelector('showFilledNumbers'));
-    const fontSizeMultiplier = useAppSelector(settingsFontSizeMultiplierSelector);
 
-    const isEmpty = sudoku.isBlankCell(cell);
     const cellBackgroundColor = getCellBgColor(
         theme,
         isActiveValue,
@@ -152,8 +90,6 @@ export const FieldCell = (props: Props) => {
         showIdenticalNumbers,
         showFilledNumbers
     );
-    const candidates = isEmpty && hasCandidates ? sudoku.getCellCandidates(cell) : [];
-
     const animation = useDerivedValue(() => withTiming(isActive ? 1 : 0, animationConfig));
 
     const cellAnimatedStyles = useAnimatedStyle(() => ({
@@ -174,28 +110,10 @@ export const FieldCell = (props: Props) => {
         { backgroundColor: cellBackgroundColor },
         cellAnimatedStyles
     ];
-    const textStyles = [
-        { color: getCellTextColor(theme, isActive, isEmpty, isHighlighted, isActiveValue, showAreas, showIdenticalNumbers) },
-        cs(isActive, styles.textActive),
-        cs(hasAnimation && hasTextAnimation, textAnimatedStyle),
-        { fontSize: CellFontSizeConstant * fontSizeMultiplier }
-    ];
 
     return (
         <ReanimatedPressable onPress={handlePress} style={cellStyles} testID={getCellSelector(props)}>
-            {candidates.map(candidate => (
-                <Reanimated.Text
-                    allowFontScaling={false}
-                    key={`candidate-${candidate}`}
-                    style={getCandidateTextStyles(theme, candidate, fontSizeMultiplier)}
-                >
-                    {candidate}
-                </Reanimated.Text>
-            ))}
-
-            <Reanimated.Text allowFontScaling={false} style={textStyles}>
-                {getText(isActive, isEmpty, hasCandidates, cell)}
-            </Reanimated.Text>
+            {children}
         </ReanimatedPressable>
     );
 };
